% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fixed.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Extract the fixed effects from a GAMM}
\usage{
extract_fixed(model, tibble = TRUE)
}
\arguments{
\item{model}{The mgcv model}

\item{tibble}{Return a tibble or standard default data.frame.  Default is
\code{TRUE}.}
}
\value{
A \code{data.frame} with the coefficients, standard error, and upper
and lower bounds based on the standard error.
}
\description{
When using a GAM for mixed models, we may have specific interest in the fixed
effect parameters.
}
\details{
This essentially reproduces the 'parametric' output from
\code{\link[mgcv]{summary.gam}}.
}
\examples{

library(mgcv)
library(lme4)

lmer_model <- lmer(Reaction ~ Days + (Days || Subject), data = sleepstudy)
ga_model <- gam(Reaction ~ Days + s(Subject, bs = "re") + s(Days, Subject, bs = "re"),
  data = sleepstudy,
  method = "REML"
)

fixef(lmer_model)
extract_fixed(ga_model)

}
\seealso{
\code{\link[lme4]{fixef}}
}
