% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot_gam}
\alias{plot_gam_2d}
\alias{plot_gam_3d}
\alias{plot_gam_by}
\alias{plot_gam_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{visibly}{\code{\link[visibly]{plot_gam}}, \code{\link[visibly]{plot_gam_2d}}, \code{\link[visibly]{plot_gam_3d}}, \code{\link[visibly]{plot_gam_by}}, \code{\link[visibly]{plot_gam_check}}}
}}

