% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fixed.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Extract the fixed effects from a GAMM}
\usage{
extract_fixed(model, ci_level = 0.95, digits = 3)
}
\arguments{
\item{model}{A gam or bam model}

\item{ci_level}{Level for the confidence interval. Must be between 0 and 1.}

\item{digits}{Rounding for the output.}
}
\value{
A \code{data.frame} with the coefficients, standard error, and upper
and lower bounds.
}
\description{
When using a GAM for mixed models, we may have specific interest in the fixed
effect parameters.
}
\details{
This essentially reproduces the 'parametric' output from
\code{\link[mgcv]{summary.gam}}.
}
\examples{

library(mgcv)
library(lme4)

lmer_model <- lmer(Reaction ~ Days + (Days || Subject), data = sleepstudy)
ga_model <- gam(Reaction ~ Days + s(Subject, bs = "re") + s(Days, Subject, bs = "re"),
  data = sleepstudy,
  method = "REML"
)

fixef(lmer_model)
extract_fixed(ga_model)

}
\seealso{
\code{\link[lme4]{fixef}}
}
